Fix recv_multishot / recvmsg_multishot writing to wrong SQE field
The IORING_RECV_MULTISHOT bit was being OR'd into `op_flags`, which
for recv-family ops aliases `msg_flags` in the kernel. Bit 1 there
is `MSG_PEEK` (0x2) -- so every "multishot recv" silently turned
into a peek that returned data without consuming it, and posted a
single CQE per call without the MORE flag set. Looked exactly like
one-shot recvs.

Per linux/io_uring.h, IORING_RECV_MULTISHOT lives in `sqe.ioprio`
(a u16), same field as IORING_ACCEPT_MULTISHOT.

Fix:
- Retype the constant from u32 to u16 so it can't compile in
  op_flags by accident.
- Move the OR from op_flags to ioprio in both recv_multishot and
  recvmsg_multishot.
- Strengthen the existing tests: assert ioprio has the bit set AND
  op_flags is clean. The old tests were green because the (wrong)
  bit happened to land where they were looking.

119 tests pass; clippy clean. Bump to 0.11.2.

+++ src/types/sendrecv.rs
@@ -5,7 +5,11 @@
 // part of the public API. The submodule is private; `mod.rs` controls
 // crate-level visibility via `pub(crate) use`.
 pub const IORING_ACCEPT_MULTISHOT: u16 = 1 << 0;
-pub const IORING_RECV_MULTISHOT: u32 = 1 << 1;
+// IORING_RECV_MULTISHOT lives in `sqe.ioprio` (a u16) — NOT in `op_flags`.
+// The kernel's `op_flags` field for recv aliases `msg_flags`, where bit 1
+// is `MSG_PEEK` (0x2). Setting this in op_flags would silently turn the
+// op into a peek recv. See linux/io_uring.h.
+pub const IORING_RECV_MULTISHOT: u16 = 1 << 1;
 
 // `ioprio` bits used by `SendRecvFlag`. Internal — callers reach these
 // through `Sqe::with(SendRecvFlag::…)`.
